+++ src/openloco/interop/hook.cpp
@@ -177,7 +177,7 @@ namespace openloco::interop
         {
             return;
         }
-        uint32_t hookaddress = (uint32_t)_hookTableAddress + (_hookTableOffset * HOOK_BYTE_COUNT);
+        uint32_t hookaddress = (intptr_t)_hookTableAddress + (_hookTableOffset * HOOK_BYTE_COUNT);
         uint8_t data[9];
         int32_t i = 0;
         data[i++] = 0xE9; // jmp
