+++ src/openloco/widget.cpp
@@ -45,9 +45,9 @@ namespace openloco::ui::widget
         regs.bx = width;
         regs.cx = x;
         regs.dx = y;
-        regs.esi = (uint32_t)window;
-        regs.edi = (uint32_t)dpi;
-        regs.ebp = (uint32_t)widget;
+        regs.esi = (intptr_t)window;
+        regs.edi = (intptr_t)dpi;
+        regs.ebp = (intptr_t)widget;
         call(0x004CF3EB, regs);
     }
